refactor: governance
