--- src/canisters/governance/canisterId.ts
@@ -1,5 +1,0 @@
-import { Principal } from "@dfinity/principal";
-
-const CANISTER_ID = Principal.fromText("rrkah-fqaaa-aaaaa-aaaaq-cai");
-
-export default CANISTER_ID;
