+++ src/constants/canister_ids.ts
@@ -1,0 +1,3 @@
+import { Principal } from "@dfinity/principal";
+
+export const GOVERNANCE_CANISTER_ID = Principal.fromText("rrkah-fqaaa-aaaaa-aaaaq-cai");
